VR: VR into proper ValueRelations
